pyproject.toml
[tool.poetry]
name = "reinforce_with_discrete_reward"
version = "0.1.0"
description = "A package to test REINFORCE with discrete reward for LLMs."
authors = ["Denis Ivanov <denivsemik@gmail.com>"]
readme = "README.md"
packages = [
    { include = "rwdr" }
]

[tool.poetry.scripts]
rwdr = "rwdr.__main__:main"

[tool.poetry.dependencies]
python = "^3.13"
hydra-core = "^1.3.2"
fire = "^0.7.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
